pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "picturescraper"
version = "0.1.0"
description = "Image discovery app using Openverse + Flickr with query analysis"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116.0",
  "pydantic>=2.7.0",
  "python-dotenv>=1.0.0",
  "requests>=2.32.0",
  "uvicorn>=0.32.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
